+++ .gitignore
@@ -1,5 +1,3 @@
 /node_modules
 .idea
 .vscode
-components
-partners
