+++ app/components/contacts/contacts-model.js
@@ -1,5 +1,5 @@
 export class ContactsModel {
-  title = '';
+  title = 'Contacts';
 
   getTitle = () => this.title;
 }
