+++ index.html
@@ -26,13 +26,13 @@
 
     </main>
     <footer class="footer">
-        <nav class="navbar fixed-bottom navbar-light bg-dark text-white pt-3">
+        <nav class="navbar navbar-light bg-dark text-white pt-3">
             <div class="col-sm-6 col-md-5 col-lg-2 col-lx-2 px-3">
                 <a class="navbar-brand text-white" href="#">Adress:</a>
                 <p>
                     <span>Ukraine, Dnipro</span>
                     <br>
-                    <span>Bohdana Chmelnytskuy Avenue</span>
+                    <span>Pobeda street, 40</span>
                 </p>
             </div>
             <div class="col-sm-6 col-md-5 col-lg-2 col-lx-1 px-3">
